requirements.txt
PyAutoGUI
PyDirectInput
Pillow
opencv-python
mss
numpy
bettercam
win32api
win32con